Refactors Payment Request flow to new design and to make it clearer. Creates new StyledQRCode component for better abstraction.
