+++ cardstack/src/components/QRCode/QRCode.tsx
@@ -28,8 +28,8 @@ export const QRCode = memo(
     size = 300,
     data,
     logo = CardstackLogo,
-    logoWidth = 40,
-    logoHeight = 40,
+    logoWidth = 34,
+    logoHeight = 36,
     logoBackgroundTransparent = false,
     logoBackgroundColor = '#fff',
     correctLevel = EasyQRCode.CorrectLevel.L,
