+++ cardstack/src/components/QRCode/index.ts
@@ -1,1 +1,2 @@
 export * from './QRCode';
+export * from './StyledQRCode';
