+++ cardstack/src/screens/PayMerchant/components/AmountInNativeCurrency.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { Text, TextProps } from '@cardstack/components';
+
+export const AmountInNativeCurrency = ({
+  amountWithSymbol,
+  textCenter = false,
+  ...textProps
+}: {
+  amountWithSymbol: string;
+  textCenter?: boolean;
+} & TextProps) => (
+  <Text
+    color="blueText"
+    fontFamily="OpenSans-Regular"
+    fontSize={12}
+    textAlign={textCenter ? 'center' : undefined}
+    {...textProps}
+  >
+    {amountWithSymbol}
+  </Text>
+);
