+++ cardstack/src/screens/PaymentRequest/components/index.ts
@@ -1,0 +1,3 @@
+export * from './PaymentRequestConfirmation';
+export * from './PaymentRequestFooter';
+export * from './PaymentRequestHeader';
